+++ arithmetization/src/main/java/net/consensys/linea/zktracer/module/bin/BinOperation.java
@@ -21,10 +21,12 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
 
-import com.google.common.base.Objects;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import lombok.experimental.Accessors;
 import net.consensys.linea.zktracer.bytestheta.BaseBytes;
+import net.consensys.linea.zktracer.container.ModuleOperation;
 import net.consensys.linea.zktracer.opcode.OpCode;
 import net.consensys.linea.zktracer.types.Bytes16;
 import net.consensys.linea.zktracer.types.UnsignedByte;
@@ -33,17 +35,14 @@ import org.apache.tuweni.bytes.Bytes32;
 
 @Getter
 @Accessors(fluent = true)
-public class BinOperation {
-  public BinOperation(OpCode opCode, BaseBytes arg1, BaseBytes arg2) {
-    this.opCode = opCode;
-    this.arg1 = arg1;
-    this.arg2 = arg2;
-  }
-
+@EqualsAndHashCode(onlyExplicitlyIncluded = true, callSuper = false)
+@RequiredArgsConstructor
+public class BinOperation extends ModuleOperation {
   private static final int LIMB_SIZE = 16;
-  private final OpCode opCode;
-  private final BaseBytes arg1;
-  private final BaseBytes arg2;
+
+  @EqualsAndHashCode.Include private final OpCode opCode;
+  @EqualsAndHashCode.Include private final BaseBytes arg1;
+  @EqualsAndHashCode.Include private final BaseBytes arg2;
   private List<Boolean> lastEightBits = List.of(false);
   private boolean bit4 = false;
   private int low4 = 0;
@@ -51,26 +50,16 @@ public class BinOperation {
   private int pivotThreshold = 0;
   private int pivot = 0;
 
-  @Override
-  public int hashCode() {
-    return Objects.hashCode(this.opCode, this.arg1, this.arg2);
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    final BinOperation that = (BinOperation) o;
-    return java.util.Objects.equals(opCode, that.opCode)
-        && java.util.Objects.equals(arg1, that.arg1)
-        && java.util.Objects.equals(arg2, that.arg2);
-  }
-
   private boolean isOneLineInstruction() {
     return (opCode == OpCode.BYTE || opCode == OpCode.SIGNEXTEND) && !arg1.getHigh().isZero();
   }
 
-  public int maxCt() {
+  @Override
+  protected int computeLineCount() {
+    return this.maxCt();
+  }
+
+  private int maxCt() {
     return isOneLineInstruction() ? 1 : LIMB_SIZE;
   }